chore(examples): examples swapped in wrong dir
